04_haadf_oft loops through all permutations of args!

+++ tests/04_haadf_otf.py
@@ -1,4 +1,4 @@
-import sys,os
+import sys,os,itertools
 try:
     import pyslice
 except ModuleNotFoundError:
@@ -18,15 +18,34 @@ dt=.005
 types={1:"B",2:"N"}
 a,b=2.4907733333333337,2.1570729817355123
 
-kwargCombos=[ {},{"ADF":True},
-             {"ADF":True,"loop_probes":10},
-             {"ADF":True,"use_memmap":True},
-             {"loop_probes":10,"use_memmap":True},
-             {"ADF":True,"loop_probes":10,"use_memmap":True} ]
+# each kwarg and it's options
+options = { "ADF":[False,True], "loop_probes":[False,10], "use_memmap":[False,True], "prism":[False,25], "store_full":[True,False] }
+args = list(options.keys())
+# all permutations: [0,0,0,0,0], [0,0,0,0,1], and so on
+indices = list(itertools.product([0,1],repeat=len(args)))
+kwargCombos = []
+for ijklm in indices:
+    dic = { k:options[k][n] for k,n in zip(args,ijklm) }
+    kwargCombos.append( dic )
+#print(kwargCombos)
 
-for kwargs in kwargCombos:
+#kwargCombos=[ {},{"ADF":True},
+#             {"ADF":True,"loop_probes":10},
+#             {"ADF":True,"use_memmap":True},
+#             {"loop_probes":10,"use_memmap":True},
+#            {"ADF":True,"loop_probes":10,"store_full":False},
+#             {"ADF":True,"use_memmap":True,"store_full":False},
+#             {"loop_probes":10,"use_memmap":True,"store_full":False},
+#             {"ADF":True,"loop_probes":10,"use_memmap":True},
+#             {"ADF":True,"loop_probes":10,"use_memmap":True,"store_full":False} ]
+
+for n,kwargs in enumerate(kwargCombos):
+    if not kwargs.get("ADF",False) and not kwargs.get("store_full",False): # skip nonsense combo (since ADF-in-post requires wavefunction_data returned)
+        continue
+    if n<18:
+        continue
     os.system("rm -rf psi_data")
-    print("RUNNING HAADF WITH KWARGS:",kwargs)
+    print("RUNNING ITERATION",n,"HAADF WITH KWARGS:",kwargs)
     trajectory=Loader(dump,timestep=dt,atom_mapping=types).load()                   # LOAD TRAJECTORY
     trajectory=trajectory.slice_positions([0,10*a],[0,10*b])                        # TRIM TO 10x10 UC
     trajectory=trajectory.get_random_timesteps(3,seed=5)                            # SELECT 10 "RANDOM" TIMESTEPS (use seed for reproducibility)
@@ -34,7 +53,7 @@ for kwargs in kwargCombos:
     probe_xs = np.linspace(10*a-a,10*a-3*a,14)                                      # SET UP GRID OF HAADF SCAN POINTS
     probe_ys = np.linspace(10*b-b,10*b-3*b,16)
     calculator.setup(trajectory,aperture=30,voltage_eV=100e3,sampling=.1,slice_thickness=.5,probe_xs=probe_xs,probe_ys=probe_ys,**kwargs)
-    if "ADF" in kwargs.keys():
+    if kwargs.get("ADF",False):
         exitwaves,haadf = calculator.run()                                          # RUN MULTISLICE, RETURNS HAADF OBJECT SINCE WE USED ADF=True FLAG
         ary = haadf.array
     else:
@@ -44,3 +63,5 @@ for kwargs in kwargCombos:
     haadf.plot("outputs/figs/04_haadf_otf.png")
     ary=np.asarray(ary)
     differ(ary[::4,::4],"outputs/haadf-test.npy","HAADF")
+    if kwargs.get("prism",False):
+        print("(but don't worry, we do not expect it to match)")
